Fixes the TODO snippet
The original TODO snippet is a mess, so I've made few changes to it. First of all, I don't need to change the date. Secondly, the date and author should be next to the TODO bit, to know who left the note. Next, I would like to go from left to right, not jump all over the place, so the indices have been reorganized. Lastly, original TODO snippet tries to modify the nickname with it's preceeding comma... why? That's fixed too.