Help the committer with restricted access to repo
Forking this project and writing my own snippets directly into it is the only way I currently know to go about snippet management in Vim. When I wanted to commit my first change, the access to my repo was revoked, the account wasn't authenticated and it shocked me a bit. Turned out that all I need to do is to change a url in .git/config file, but I think it would be a nice idea to keep this information in the README."